Debugging binary tree

+++ 15-binary_tree_is_full.c
@@ -5,7 +5,7 @@
  * @tree: pointer to the root node
  * Return: 1 if full, else 0
  */
-size_t binary_tree_is_full(const binary_tree_t *tree)
+int binary_tree_is_full(const binary_tree_t *tree)
 {
 	if (!tree)
 		return (0);
